[GOBBLIN-255] Move builder related methods to ParquetHdfsWriterBuilder
