+++ gobblin-core/src/main/java/org/apache/gobblin/writer/ParquetHdfsDataWriter.java
@@ -17,27 +17,13 @@
 package org.apache.gobblin.writer;
 
 import java.io.IOException;
-import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 
 import org.apache.gobblin.configuration.ConfigurationKeys;
 import org.apache.gobblin.configuration.State;
-import org.apache.gobblin.util.ForkOperatorUtils;
-import org.apache.hadoop.conf.Configuration;
 
-import parquet.column.ParquetProperties;
 import parquet.example.data.Group;
 import parquet.hadoop.ParquetWriter;
-import parquet.hadoop.example.GroupWriteSupport;
-import parquet.hadoop.metadata.CompressionCodecName;
-import parquet.schema.MessageType;
-
-import static org.apache.gobblin.configuration.ConfigurationKeys.WRITER_CODEC_TYPE;
-import static org.apache.gobblin.writer.ParquetDataWriterBuilder.*;
-import static parquet.hadoop.ParquetWriter.DEFAULT_BLOCK_SIZE;
-import static parquet.hadoop.ParquetWriter.DEFAULT_IS_DICTIONARY_ENABLED;
-import static parquet.hadoop.ParquetWriter.DEFAULT_IS_VALIDATING_ENABLED;
-import static parquet.hadoop.ParquetWriter.DEFAULT_PAGE_SIZE;
 
 
 /**
@@ -51,25 +37,14 @@ import static parquet.hadoop.ParquetWriter.DEFAULT_PAGE_SIZE;
  * @author tilakpatidar
  */
 public class ParquetHdfsDataWriter extends FsDataWriter<Group> {
-  private final MessageType schema;
   private final ParquetWriter<Group> writer;
   protected final AtomicLong count = new AtomicLong(0);
-  private final long pageSize;
-  private final long dictPageSize;
-  private final boolean enableDictionary;
-  private final boolean validate;
   public static final String DEFAULT_PARQUET_WRITER = "v1";
 
-  public ParquetHdfsDataWriter(FsDataWriterBuilder<MessageType, Group> builder, State state)
+  public ParquetHdfsDataWriter(ParquetDataWriterBuilder builder, State state)
       throws IOException {
     super(builder, state);
-    this.schema = builder.getSchema();
-    this.pageSize = state.getPropAsLong(getProperty(WRITER_PARQUET_PAGE_SIZE), DEFAULT_PAGE_SIZE);
-    this.dictPageSize = state.getPropAsLong(getProperty(WRITER_PARQUET_DICTIONARY_PAGE_SIZE), DEFAULT_BLOCK_SIZE);
-    this.enableDictionary =
-        state.getPropAsBoolean(getProperty(WRITER_PARQUET_DICTIONARY), DEFAULT_IS_DICTIONARY_ENABLED);
-    this.validate = state.getPropAsBoolean(getProperty(WRITER_PARQUET_VALIDATE), DEFAULT_IS_VALIDATING_ENABLED);
-    this.writer = buildParquetWriter();
+    this.writer = builder.getWriter(this.blockSize, this.stagingFile);
   }
 
   @Override
@@ -93,27 +68,5 @@ public class ParquetHdfsDataWriter extends FsDataWriter<Group> {
       throws IOException {
     this.writer.close();
     super.close();
-  }
-
-  private ParquetWriter<Group> buildParquetWriter()
-      throws IOException {
-    CompressionCodecName codec = getCodecFromConfig();
-    GroupWriteSupport support = new GroupWriteSupport();
-    Configuration conf = new Configuration();
-    GroupWriteSupport.setSchema(this.schema, conf);
-    ParquetProperties.WriterVersion writerVersion = ParquetProperties.WriterVersion
-        .fromString(this.properties.getProp(getProperty(WRITER_PARQUET_VERSION), DEFAULT_PARQUET_WRITER));
-    return new ParquetWriter<>(this.stagingFile, support, codec, (int) this.blockSize, (int) this.pageSize,
-        (int) this.dictPageSize, this.enableDictionary, this.validate, writerVersion, conf);
-  }
-
-  private CompressionCodecName getCodecFromConfig() {
-    String codecValue = Optional.ofNullable(this.properties.getProp(getProperty(WRITER_CODEC_TYPE)))
-        .orElse(CompressionCodecName.SNAPPY.toString());
-    return CompressionCodecName.valueOf(codecValue.toUpperCase());
-  }
-
-  private String getProperty(String key) {
-    return ForkOperatorUtils.getPropertyNameForBranch(key, this.numBranches, this.branchId);
   }
 }
